+++ internal/scraper/india/cii.go
@@ -12,7 +12,7 @@ import (
 
 func GetCostInflationIndex() ([]*cii.CII, error) {
 	log.Info("Fetching Cost Inflation Index from Purified Bytes")
-	resp, err := http.Get("https://india.purifiedbytes.com/api/cii/v2.json")
+	resp, err := http.Get("https://india.finbodhi.com/api/cii/v2.json")
 	if err != nil {
 		return nil, err
 	}